+++ CHANGELOG.md
@@ -7,6 +7,23 @@
 - Add User Interaction Tracing for Touch events ([#2835](https://github.com/getsentry/sentry-react-native/pull/2835))
 - Add Gesture Tracing for React Native Gesture Handler API v2 ([#2865](https://github.com/getsentry/sentry-react-native/pull/2865))
 
+### Fixes
+
+- Fix use Fetch transport when option `enableNative` is `false` ([#2897](https://github.com/getsentry/sentry-react-native/pull/2897))
+- Improve logs when `enableNative` is `false` ([#2897](https://github.com/getsentry/sentry-react-native/pull/2897))
+
+### Dependencies
+
+- Bump JavaScript SDK from v7.40.0 to v7.44.0 ([#2874](https://github.com/getsentry/sentry-react-native/pull/2874), [#2908](https://github.com/getsentry/sentry-react-native/pull/2908))
+  - [changelog](https://github.com/getsentry/sentry-javascript/blob/develop/CHANGELOG.md#7440)
+  - [diff](https://github.com/getsentry/sentry-javascript/compare/7.40.0...7.44.0)
+
+## 5.1.1
+
+### Fixes
+
+- Remove non URL `frame.abs_path` which was causing source maps to fail ([#2891](https://github.com/getsentry/sentry-react-native/pull/2891))
+
 ### Dependencies
 
 - Bump Cocoa SDK from v8.2.0 to v8.3.0 ([#2876](https://github.com/getsentry/sentry-react-native/pull/2876))
@@ -15,9 +32,6 @@
 - Bump CLI from v2.14.3 to v2.14.4 ([#2873](https://github.com/getsentry/sentry-react-native/pull/2873))
   - [changelog](https://github.com/getsentry/sentry-cli/blob/master/CHANGELOG.md#2144)
   - [diff](https://github.com/getsentry/sentry-cli/compare/2.14.3...2.14.4)
-- Bump JavaScript SDK from v7.40.0 to v7.42.0 ([#2874](https://github.com/getsentry/sentry-react-native/pull/2874))
-  - [changelog](https://github.com/getsentry/sentry-javascript/blob/develop/CHANGELOG.md#7420)
-  - [diff](https://github.com/getsentry/sentry-javascript/compare/7.40.0...7.42.0)
 
 ## 5.1.0